extract CollectionEditMenu
